Kicad diagram for the seat PCB of the V2 chair
Includes the whole project folder under the Diagram folder.

A symbol library has been created to hold a modifed version of the MCP3008 symbol from the standard KiCad library.
This version has a Dout pint with the tri-state electrical mode, instead of the output mode.
This stops the ERC from raising errors when multiple Dout pins are connected by the SPI bus.

An extra footprint library was added to provide footprints of the right-angle Molex 2.54mm pitch KK connectors.
The standard library only has the vertical version.
Footprints are yet to be assigned to the schematic.